rename image files
