+++ src/assets/images/avatar.png
